Simplify build environment initialization process
