+++ pyodide-build/pyodide_build/cli/build_recipes.py
@@ -2,10 +2,9 @@ from pathlib import Path
 
 import typer
 
-from .. import buildall, buildpkg, common, pywasmcross
-from ..common import get_num_cores, init_environment
+from .. import build_env, buildall, buildpkg, common, pywasmcross
+from ..common import get_num_cores
 from ..logger import logger
-from ..out_of_tree.utils import initialize_pyodide_root
 
 
 def recipe(
@@ -67,9 +66,9 @@ def recipe(
     ),
 ) -> None:
     """Build packages using yaml recipes and create repodata.json"""
-    initialize_pyodide_root()
+    build_env.init_environment()
 
-    if common.in_xbuildenv():
+    if build_env.in_xbuildenv():
         common.check_emscripten_version()
 
     root = Path.cwd()
@@ -80,8 +79,6 @@ def recipe(
 
     if not recipe_dir_.is_dir():
         raise FileNotFoundError(f"Recipe directory {recipe_dir_} not found")
-
-    init_environment()
 
     build_args = pywasmcross.BuildArgs(
         cflags=cflags,
