+++ pyodide-build/pyodide_build/cli/config.py
@@ -1,7 +1,6 @@
 import typer
 
-from ..common import get_make_environment_vars
-from ..out_of_tree.utils import initialize_pyodide_root
+from ..build_env import get_build_environment_vars, init_environment
 
 app = typer.Typer(help="Manage config variables used in pyodide")
 
@@ -19,9 +18,7 @@ def callback() -> None:
 
 
 def _get_configs() -> dict[str, str]:
-    initialize_pyodide_root(quiet=True)
-
-    configs: dict[str, str] = get_make_environment_vars()
+    configs: dict[str, str] = get_build_environment_vars()
 
     configs_filtered = {k: configs[v] for k, v in PYODIDE_CONFIGS.items()}
     return configs_filtered
@@ -47,6 +44,8 @@ def get_config(
     """
     Get a value of a single config variable used in pyodide
     """
+    init_environment(quiet=True)
+
     configs = _get_configs()
 
     if config_var not in configs:
