+++ pyodide-build/pyodide_build/out_of_tree/build.py
@@ -2,7 +2,7 @@ import os
 from pathlib import Path
 from typing import Any
 
-from .. import common, pypabuild
+from .. import build_env, common, pypabuild
 
 
 def run(exports: Any, args: list[str], outdir: Path | None = None) -> Path:
@@ -14,8 +14,9 @@ def run(exports: Any, args: list[str], outdir: Path | None = None) -> Path:
     cxxflags += f" {os.environ.get('CXXFLAGS', '')}"
     ldflags = common.get_make_flag("SIDE_MODULE_LDFLAGS")
     ldflags += f" {os.environ.get('LDFLAGS', '')}"
+
     env = os.environ.copy()
-    common.set_build_environment(env)
+    env.update(build_env.get_build_environment_vars())
 
     build_env_ctx = pypabuild.get_build_env(
         env=env,
